refactor: 메시지 관련 패키지 구조 변경
메시지 enum 관련 패키지를 분리합니다.
줄바꿈 출력을 입력과 동시에 하도록 변경합니다.
