+++ src/main/java/lotto/view/InputView.java
@@ -6,7 +6,7 @@ public class InputView {
 
     private static final String REQUEST_PURCHASE_AMOUNT_MESSAGE = "구입금액을 입력해 주세요.";
     private static final String REQUEST_WINNING_NUMBERS_MESSAGE = "당첨 번호를 입력해 주세요.";
-    private static final String REQUEST_BONUS_NUMBER_MESSAGE = "\n보너스 번호를 입력해 주세요.";
+    private static final String REQUEST_BONUS_NUMBER_MESSAGE = "보너스 번호를 입력해 주세요.";
 
     public String requestPurchaseAmount() {
         System.out.println(REQUEST_PURCHASE_AMOUNT_MESSAGE);
@@ -25,6 +25,7 @@ public class InputView {
 
     private String readUserInput() {
         String input = Console.readLine();
+        System.out.println();
         return input.trim();
     }
 }
