+++ src/main/java/lotto/view/OutputView.java
@@ -2,30 +2,32 @@ package lotto.view;
 
 import java.util.List;
 import lotto.dto.WinningResult;
-import lotto.util.message.WinningMatchMessage;
+import lotto.message.WinningMatchMessage;
 
 public class OutputView {
 
-    private static final String PURCHASED_QUANTITY_MESSAGE = "\n%d개를 구매했습니다.\n";
-    private static final String WINNING_RESULT_TITLE_MESSAGE = "\n당첨 통계\n---\n%s";
-    private static final String TOTAL_YIELD_MESSAGE = "총 수익률은 %.1f%%입니다.%n";
-    private static final String COUNT_MESSAGE = " - %d개\n";
+    private static final String PURCHASED_QUANTITY_MESSAGE = "%d개를 구매했습니다.";
+    private static final String WINNING_RESULT_TITLE_MESSAGE = "당첨 통계%n---%n%s";
+    private static final String COUNT_MESSAGE = " - %d개%n";
+    private static final String TOTAL_YIELD_MESSAGE = "총 수익률은 %.1f%%입니다.";
 
     public void printPurchasedQuantity(int purchasedQuantity) {
-        System.out.printf(PURCHASED_QUANTITY_MESSAGE, purchasedQuantity);
+        String message = String.format(PURCHASED_QUANTITY_MESSAGE, purchasedQuantity);
+        System.out.println(message);
     }
 
     public void printLottoTickets(List<List<Integer>> lottoTickets) {
-        StringBuilder stringBuilder = new StringBuilder();
+        StringBuilder message = new StringBuilder();
         for (List<Integer> lottoTicket : lottoTickets) {
-            stringBuilder.append(lottoTicket.toString())
+            message.append(lottoTicket.toString())
                     .append(System.lineSeparator());
         }
-        System.out.println(stringBuilder);
+        System.out.println(message);
     }
 
     public void printWinningResult(WinningResult winningResult) {
-        System.out.printf(WINNING_RESULT_TITLE_MESSAGE, buildWinningStatistics(winningResult));
+        String message = String.format(WINNING_RESULT_TITLE_MESSAGE, buildWinningStatistics(winningResult));
+        System.out.println(message);
     }
 
     private String buildWinningStatistics(WinningResult winningResult) {
@@ -39,10 +41,14 @@ public class OutputView {
                 + getResultCount(winningResult.fiveWithBonusCount())
                 + WinningMatchMessage.FIRST.get()
                 + getResultCount(winningResult.sixMatchesCount())
-                + String.format(TOTAL_YIELD_MESSAGE, winningResult.totalYield());
+                + getTotalYield(winningResult);
     }
 
     private String getResultCount(int count) {
         return String.format(COUNT_MESSAGE, count);
+    }
+
+    private String getTotalYield(WinningResult winningResult) {
+        return String.format(TOTAL_YIELD_MESSAGE, winningResult.totalYield());
     }
 }
