+++ src/test/java/lotto/ApplicationTest.java
@@ -199,23 +199,23 @@ class ApplicationTest extends NsTest {
     @Test
     void WhenAllReTry() {
         assertSimpleTest(() -> run(
-                "999",
-                "1001",
-                "abc",
-                "1000",
+                        "999",
+                        "1001",
+                        "abc",
+                        "1000",
 
-                "1,2,3,4,5,a",
-                "1,2,3,4,5",
-                "1,2,3,4,5,6,7",
-                "1,2,3,4,5,46",
-                "1,2,3,4,5,5",
-                "1,2,3,4,5,6",
+                        "1,2,3,4,5,a",
+                        "1,2,3,4,5",
+                        "1,2,3,4,5,6,7",
+                        "1,2,3,4,5,46",
+                        "1,2,3,4,5,5",
+                        "1,2,3,4,5,6",
 
-                "a",
-                "0",
-                "6",
-                "7"
-        )
+                        "a",
+                        "0",
+                        "6",
+                        "7"
+                )
         );
     }
