+++ src/test/java/lotto/model/shop/LottoMachineTest.java
@@ -3,7 +3,7 @@ package lotto.model.shop;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.*;
 
-import lotto.util.message.LottoErrorMessage;
+import lotto.message.LottoErrorMessage;
 import lotto.model.ticket.LottoTickets;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
