+++ src/test/java/lotto/model/win/BonusNumberTest.java
@@ -3,7 +3,7 @@ package lotto.model.win;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import lotto.util.message.LottoErrorMessage;
+import lotto.message.LottoErrorMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
