+++ src/test/java/lotto/model/win/WinningNumbersTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.*;
 
 import java.util.List;
-import lotto.util.message.LottoErrorMessage;
+import lotto.message.LottoErrorMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
