+++ src/test/java/lotto/util/InputParserTest.java
@@ -3,7 +3,7 @@ package lotto.util;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import lotto.util.message.InputErrorMessage;
+import lotto.message.InputErrorMessage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
